backend/pyproject.toml
[project]
name = "backend-Glyph"
version = "0.1.0"
description = "the backend for the glyph application"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.42.39",
    "botocore>=1.42.39",
    "flask>=3.1.2",
    "flask-cors>=6.0.2",
    "flask-sqlalchemy>=3.1.1",
    "google-genai>=1.57.0",
    "graphviz>=0.21",
    "huggingface-hub>=0.36.0",
    "langchain>=1.2.3",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.1.3",
    "langchain-huggingface>=1.2.0",
    "marshmallow>=3.26.2",
    "pdf2image>=1.17.0",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pyjwt>=2.10.1",
    "pypdf>=6.6.0",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.45",
]
